Major changes to CrIS rtp creation.
First shot on CrIS rtp using new routines for converting
SARTA IASI to CrIS radiances.  Now using Howard's iasi2rtp.m

I am a little worried about the use of guard_ind.m in
ccast2rtp.m.  Everything seems to work, but I think guard_ind.m is
wrong, but is "corrected" in ccast2rtp.m

+++ emis/emis_danz.m
@@ -20,8 +20,7 @@ newc = zeros(10,nobs);
 
 % Now figure out numerical months
 mtime = datetime(1993,1,1,0,0,rtime);
-mon = day(mtime,'dayofyear')/365;
-
+mon = single(day(mtime,'dayofyear')/365);
 % Find interpolated expansion coefficients (linear is default)
 for i=1:10
    newc(i,:) = danz(i).emis(lon,lat,mon);
